Accidently wrote code into the wrong component, copied code into the correct component
