+++ water-plants/src/App.js
@@ -1,11 +1,11 @@
 import React from 'react';
 import './App.css';
-import LoginForm from "./components/LoginForm"
+import SignUpForm from "./components/SignUpForm";
 
 function App() {
   return (
     <div className="App">
-      <LoginForm/>
+      <SignUpForm/>
     </div>
   );
 }
